+++ BTrees/BTreeModuleTemplate.c
@@ -77,7 +77,25 @@ static void PyVar_Assign(PyObject **v, PyObject *e) { Py_XDECREF(*v); *v=e;}
 #error "PY_LONG_LONG required but not defined"
 #endif
 
+static int
+longlong_handle_overflow(PY_LONG_LONG result, int overflow)
+{
+    if (overflow)
+    {
+        /* Python 3 tends to have an exception already set, Python 2 not so much */
+        if (!PyErr_Occurred())
+            PyErr_SetString(PyExc_OverflowError, "couldn't convert integer to C long long");
+        return 0;
+    }
+    else if (result == -1 && PyErr_Occurred())
+        /* An exception has already been raised. */
+        return 0;
+    return 1;
+}
+
+
 #ifdef NEED_LONG_LONG_KEYS
+
 #if defined(ZODB_UNSIGNED_VALUE_INTS) || defined(ZODB_UNSIGNED_KEY_INTS)
 static int
 ulonglong_check(PyObject *ob)
@@ -85,21 +103,25 @@ ulonglong_check(PyObject *ob)
 #ifndef PY3K
     if (PyInt_Check(ob))
     {
-	long tmp;
-	tmp = PyInt_AS_LONG(ob);
-	if (tmp < 0) {
-	    PyErr_SetString(PyExc_OverflowError, "unsigned value less than 0");
-	    return 0;
-	}
-	return 1;
+        long tmp;
+        tmp = PyInt_AS_LONG(ob);
+        if (tmp < 0) {
+            PyErr_SetString(PyExc_OverflowError, "unsigned value less than 0");
+            return 0;
+        }
+        return 1;
     }
 #endif
 
     if (!PyLong_Check(ob))
-	return 0;
+    {
+        return 0;
+    }
 
     if (PyLong_AsUnsignedLongLong(ob) == (unsigned long long)-1 && PyErr_Occurred())
-	return 0;
+    {
+        return 0;
+    }
     return 1;
 }
 #endif /* defined(ZODB_UNSIGNED_VALUE_INTS) || defined(ZODB_UNSIGNED_KEY_INTS) */
@@ -115,15 +137,10 @@ longlong_check(PyObject *ob)
 
     if (PyLong_Check(ob)) {
         int overflow;
-        (void)PyLong_AsLongLongAndOverflow(ob, &overflow);
-        if (overflow)
-            goto overflow;
-        return 1;
+        PY_LONG_LONG result;
+        result = PyLong_AsLongLongAndOverflow(ob, &overflow);
+        return longlong_handle_overflow(result, overflow);
     }
-    return 0;
-overflow:
-    PyErr_SetString(PyExc_ValueError,
-                    "longlong_check: long integer out of range");
     return 0;
 }
 
@@ -141,15 +158,17 @@ ulonglong_as_object(unsigned PY_LONG_LONG val)
 static int
 ulonglong_convert(PyObject *ob, unsigned PY_LONG_LONG *value)
 {
+    unsigned PY_LONG_LONG val;
+
 #ifndef PY3K
     if (PyInt_Check(ob))
     {
-	long tmp;
-	tmp = PyInt_AS_LONG(ob);
-	if (tmp < 0) {
-	    PyErr_SetString(PyExc_OverflowError, "unsigned value less than 0");
-	    return 0;
-	}
+        long tmp;
+        tmp = PyInt_AS_LONG(ob);
+        if (tmp < 0) {
+            PyErr_SetString(PyExc_OverflowError, "unsigned value less than 0");
+            return 0;
+        }
         (*value) = (unsigned PY_LONG_LONG)tmp;
         return 1;
     }
@@ -161,10 +180,9 @@ ulonglong_convert(PyObject *ob, unsigned PY_LONG_LONG *value)
         return 0;
     }
 
-    unsigned PY_LONG_LONG val;
     val = PyLong_AsUnsignedLongLong(ob);
     if (val == (unsigned long long)-1 && PyErr_Occurred())
-	return 0;
+        return 0;
     (*value) = val;
     return 1;
 }
@@ -181,6 +199,8 @@ longlong_as_object(PY_LONG_LONG val)
 static int
 longlong_convert(PyObject *ob, PY_LONG_LONG *value)
 {
+    PY_LONG_LONG val;
+    int overflow;
 #ifndef PY3K
     if (PyInt_Check(ob))
     {
@@ -194,19 +214,13 @@ longlong_convert(PyObject *ob, PY_LONG_LONG *value)
         PyErr_SetString(PyExc_TypeError, "expected integer key");
         return 0;
     }
-    else
+    val = PyLong_AsLongLongAndOverflow(ob, &overflow);
+    if (!longlong_handle_overflow(val, overflow))
     {
-        PY_LONG_LONG val;
-        int overflow;
-        val = PyLong_AsLongLongAndOverflow(ob, &overflow);
-        if (overflow)
-            goto overflow;
-        (*value) = val;
-        return 1;
+        return 0;
     }
-overflow:
-    PyErr_SetString(PyExc_ValueError, "long integer out of range");
-    return 0;
+    (*value) = val;
+    return 1;
 }
 
 #endif  /* NEED_LONG_LONG_SUPPORT */