+++ BTrees/tests/testBTrees.py
@@ -443,7 +443,17 @@ class FamilyTest(unittest.TestCase):
         self.assertEqual(
             BTrees.family32.IF, BTrees.IFBTree)
         self.assertEqual(
+            BTrees.family32.UO, BTrees.UOBTree)
+        self.assertEqual(
+            BTrees.family32.OU, BTrees.OUBTree)
+        self.assertEqual(
+            BTrees.family32.UU, BTrees.UUBTree)
+        self.assertEqual(
+            BTrees.family32.UF, BTrees.UFBTree)
+        self.assertEqual(
             BTrees.family32.OO, BTrees.OOBTree)
+        self.assertEqual(
+            BTrees.family32.OU, BTrees.OUBTree)
         s = IOTreeSet()
         s.insert(BTrees.family32.maxint)
         self.assertTrue(BTrees.family32.maxint in s)
@@ -475,7 +485,17 @@ class FamilyTest(unittest.TestCase):
         self.assertEqual(
             BTrees.family64.IF, BTrees.LFBTree)
         self.assertEqual(
+            BTrees.family64.UO, BTrees.QOBTree)
+        self.assertEqual(
+            BTrees.family64.OU, BTrees.OQBTree)
+        self.assertEqual(
+            BTrees.family64.UU, BTrees.QQBTree)
+        self.assertEqual(
+            BTrees.family64.UF, BTrees.QFBTree)
+        self.assertEqual(
             BTrees.family64.OO, BTrees.OOBTree)
+        self.assertEqual(
+            BTrees.family64.OU, BTrees.OQBTree)
         s = LOTreeSet()
         s.insert(BTrees.family64.maxint)
         self.assertTrue(BTrees.family64.maxint in s)