+++ CHANGES.rst
@@ -13,6 +13,13 @@
 - Fix the value for ``BTrees.OIBTree.using64bits`` when using the pure Python
   implementation (PyPy and when ``PURE_PYTHON`` is in the environment).
 
+- Make the errors that are raised when values are out of range more
+  consistent between Python 2 and Python 3 and between 32-bit and
+  64-bit variants.
+
+- Fix the value for ``BTrees.OIBTree`` when using the pure Python
+  implementation (PyPy and when ``PURE_PYTHON`` is in the environment).
+
 4.6.1 (2019-11-07)
 ==================